Modifico como dar de alta los puntos v1
